% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookclubrunner-package.R
\docType{package}
\name{bookclubrunner-package}
\alias{bookclubrunner}
\alias{bookclubrunner-package}
\title{bookclubrunner: R4DS Book Club Management}
\description{
Automate or otherwise streamline tasks associated with running R4DS book clubs, such as choosing the best time for a new cohort.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/bookclubrunner}
  \item \url{https://r4ds.github.io/bookclubrunner/}
  \item Report bugs at \url{https://github.com/r4ds/bookclubrunner/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
